Rename to ROM::EncryptedAttribute
This was suggested by Piotr Solnica on ROM
forum (https://discourse.rom-rb.org/t/encrypted-attributes/637/2)
so I'm following this convention.
