+++ lib/rom/encrypted_attribute/version.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module ROM
+  module EncryptedAttribute
+    VERSION = "0.0.3"
+  end
+end
