--- lib/rom_encrypted_attribute/version.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module RomEncryptedAttribute
-  VERSION = "0.0.3"
-end
